Adding changes required for this code to be used in IDC Inflow

+++ base/5_env_dev_utils_setup.sh
@@ -89,7 +89,7 @@ EOF
 
 # enable and start the service
 sudo systemctl enable set-cpufreq-governor.service
-sudo systemctl start set-cpufreq-governor.service
+[ -z "$OMMIT_SERVICE_START" ] && sudo systemctl start set-cpufreq-governor.service
 
 # install and setup Docker
 if $INSTALL_DOCKER; then
@@ -118,6 +118,10 @@ echo "ulimit -u 65535" | sudo tee -a /etc/security/limits.conf
 # inform user
 colored_output "Cleanup..." blue
 sudo apt -y autoremove
-colored_output "Setup completed. Rebooting in 10 seconds..." blue
-sleep 10
-sudo reboot
+
+if [ -z "$OMMIT_REBOOT" ]
+then
+    colored_output "Setup completed. Rebooting in 10 seconds..." blue
+    sleep 10
+    sudo reboot
+fi
